+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 <a name="Pending Release"></a>
 ## [Pending Release](https://github.com/lightstep/otel-launcher-java/compare/0.8.1...master)
 * Replaced lightstep.hostname tag with host.name
+* Use OpenTelemetry Java API version 0.9.1
+* Use OpenTelemetry Java Instrumentation version 0.9.0
 
 <a name="0.8.1"></a>
 ## [0.8.1](https://github.com/lightstep/otel-launcher-java/compare/0.8.0...0.8.1)